Refactor unit test driver as basis for feature test driver
TraceurTestRunner extends Mocha6 and contains common code for
unit and feature tests.  unitTestRunner is the instance for
unit tests.  The module that runs the unit tests is now
runUnitTests.js.

Rename testRunner to unitTestRunner.
