+++ test/test-utils.js
@@ -95,7 +95,7 @@
 
   // Replace the Closure-provided array comparer with our own that doesn't barf
   // because Array.prototype has a __iterator__ method.
-  // TODO(jjb): get this from import on testRunner
+  // TODO(jjb): get this from import on unitTests
   function assertArrayEquals(expected, actual) {
     assert.equal(JSON.stringify(actual, null, 2),
                  JSON.stringify(expected, null, 2));
