+++ test/unit/codegeneration/PureES6Transformer.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {suite, test, assert} from '../../modular/testRunner.js';
+import {suite, test, assert} from '../../unit/unitTestRunner.js';
 
 import {ErrorReporter} from '../../../src/util/ErrorReporter.js';
 import {Options} from '../../../src/Options.js';
