+++ test/unit/util/ErrorReporter.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {suite, test, setup, teardown, assert} from '../../modular/testRunner.js';
+import {suite, test, setup, teardown, assert} from '../../unit/unitTestRunner.js';
 
 suite('ErrorReporter.js', function() {
